Rename CLI arguments for disambiguation
There will be ports used on both the p2p endpoint and the json-rpc
endpoint. It's too ambiguous to have a CLI argument of simply "http
port".

Also, single-letter flags are too short to handle the argument
complexity that's coming. Not to mention that we should not overwrite
the -h (help) arg.