PortalStorage Struct for abstracting Trin storage logic (#69)
This is an implementation of a PortalStorage struct that handles the logic of determining our node's radius at any given time, and determining if/how to store a given piece of data given our node's current radius and data capacity.

The functions that it provides to the rest of the client code are:

new(config: PortalStorageConfig) -> Result<(), PortalStorageError>: The constructor takes a PortalStorageConfig, which allows user to specify the client's max capacity in kb
should_store(key: String) -> Result<bool, PortalStorageError> : used for determining whether to respond with an ACCEPT to an OFFER request
store(key: String, value: String) -> Result<(), PortalStorageError> for storing a new piece of content from a STORE request. This function automatically takes into account the capacity and, if necessary, drops the farthest piece of data, finds the new farthest and updates the radius.
get(key: String) -> Result<(), PortalStorageError>: Gets the value for the key if it exists in our DB.
get_current_radius() -> U256: Gets the current radius (currently u64, should probably be a u256 because I'm under the assumption that a radius of infinity is going to be represented over the wire by u256::max).
get_total_storage_usage_in_bytes_on_disk() -> Result<u64, PortalStorageError>: Gets size of Portal Network data (including storage overhead) stored on disk.
Notes:

This wraps RocksDB as the main keystore, but is also using SQLite in order to calculate the xor-farthest key on the fly (once we drop the old farthest key). In the code I refer to this SQLite DB as the "meta_db".

This also includes the generate-data bin, which can be run as so: cargo run --bin generate-data -- --kb 120, which will populate the RocksDB instance with 120kb of random key value pairs. Use cargo run --bin generate-data -- -h to see all options. I plan to use this to help while testing that the intricacies of the store function are working properly. I also plan to add get and delete functionality, so it will probably end up being called trin-db-util instead of generate-data.