Faster Windows builds in CircleCI
We already test the build implicitly when running the test, no need for
a separate build that takes 6 minutes.

We already run clippy in the Linux environment. It's not worth another 6
minutes on each build to run it again in Windows.

There was a hack in the Clippy installer to get cargo git checkouts
working in CircleCI. Since that's removed, we revisited the hack, and
the cleaner workaround of using .cargo/config was put in place.

For more, see: https://github.com/rust-lang/cargo/issues/2078