Reduce flakiness of test_offer_accept()
Since portal_historySendOffer returns before the peer accepts and stores
content, there is a race condition where the peer has not finished
storing the offered content at the moment the test is validating the
peer's database. So give the peer some extra time, if needed, to finish
out the offer/accept flow.