Keep data radius only in ContentStore
There were duplicates in overlay and overlay_service that were not being
updating as the storage limit was exceeded. This meant that trin was
always using the user-configured (MAX) radius for all ping/pong
messages.

It meant we were accepting lots of data outside our radius, and then
complaining about receiving them in the logs.