Only use ANSI colors if output to colored terminal
Only import the libraries as needed, per OS.

Unfortunately, increases the size of trin-utlis a bit, to add log tracing.

On *nix, check if output is going to terminal (or for example, piped to
a file). Also, if going to a terminal, make sure the terminal supports
color (ie~ the environment variable TERM is not "dumb").