Only prune until DB is slightly overfull
Previously, we pruned until it was slightly underfull.

Add new test for what happens when we restart with same storage limit.
It should correctly detect the radius. (This is one of the biggest
benefits of overfilling the DB a bit)

Separately, we update the decreasing-storage test for our new behavior.
Our new behavior means one more item is kept than previously expected.