Radius metric: higher precision & switch to ratio
Get more bytes of the radius, to get a higher precision estimate. We
don't expect to display a higher precision than 1 in a million, so 3
bytes would be sufficient. The nearest supported integer is at u32 with
4 bytes.

Ratio (0-1) is favored over % (0-100) in the prometheus system. Switch
to saving that value, plus the corresponding metric name change.