Remove old broken dashboard, and json-rpc plugin
The old dashboard is broken. Neither Mike nor I felt like digging into
why, so just decided to remove it instead.

The job of the plugin was to provide json-rpc support for some panels in
that dashboard. Since that feature isn't required anymore, the plugin
was removed, and the datasource is no longer created.