refactor: combine duplicate code to send utp data
There was already a decent amount of drift: only one location was
calling shutdown(), the metrics were handled differently, etc. It seems
like we want to handle all these things the same in both places.

This came up when Nick pointed out that there was a second utp-handling
code chunk in overlay_service.