chore: reth-ipc to 0.1.0-alpha.6, json to v0.20.0
Upgrade the reth and jsonrpsee versions everywhere in the workspace.
Re-build the lockfile with both upgrades.

Upgrading reth-ipc requires a jsonpsee upgrade. These two updates have
to be combined, otherwise the two different jsonrpsee versions (one for
reth-ipc, the other for the rest of trin) cause their own build problems.

Handle the several breaking changes introduced since ~v0.16 of
jsonrpsee, most notably:
- Error handling changed a lot -- there's now a preference to build your
  own custom error with a translation to their ErrorObject trait.
- server.start() is infallible now.
- The server protocols are combined, so it's just a Server instead of an
  HttpServer.
- A few classes moved around.

The reth-ipc version change like a downgrade, but that's just how the
version numbers changed. reth-ipc used to be v0.1.0, and now it's a
pre-release (because they linked the versions of every crate in the repo
together).