refactor(bridge): Mutex & Atomic to value-passing
We can skip the locks and value-wrapping by simply passing in the
variables we care about, and returning it to use at the start of the
next loop.

Also, there's a spawn that's not necessary anymore, because we were
waiting for the result anyway.

The full data service should happen within the event loop. If it
continuously takes longer than the slot length, then the bridge will
fall further and further behind mainnet.