feat: add base eth_* rpc & eth_chainID is always 1
This is part of the basic structure required before adding a more
complex endpoint like eth_getBlockByHash

Add a test that eth_chainId is recognized by:
- our own ethportal-api RPC client
- an unmodified 3rd party RPC client (ethers-rs)

This helps test that we built our own client correctly, and that our
server is following standards. For example, the chainID was being sent
as an integer before, instead of hex-encoded. ethers-rs helped us detect
this issue.

The new tests/rpc_server.rs is the home for testing our server with the
3rd-party ethers-rs web3 client.