perf(light-client): faster builds w/ no ethers-rs
After the latest slimming-down of light-client, we could remove ethers
completely. When we eventually add the functionality back in, we are
likely to use internal types instead of the ones from ethers anyway, so
this is probably a permanent-ish change.

I was noticing that a recent PR that added ethers 2.0 for testing the
RPC server was causing CI to crash while building. It seemed like it was
the Out of Memory Killer that sigkilled the build. Ethers was the only
big thing added, so I stripped it down, and that helped some, but while
I was getting familiar with ethers, I thought I'd also do this change to
make the build even a little bit smaller and faster.

The problem was multiplied because light-client was using a different
version of ethers, so two copies were built.