fix: better utp connection rates w/ default config
We were seeing failed connection attempts fairly regularly on the
testnet. This is one step of resolving that issue.

utp-rs is mostly tested using the default configurations, which have
changed a few times to improve results. As far as I can tell, the
configuration choices here are just an old copy of the config.

So instead of copying in new values, I switched the config to use
defaults whenever possible. (which requires using lazy_static)
The only custom configuration that seems to matter is that discv5
enforces a custom maximum packet size.

Some reasons I think this will help:
- Number of connection attempts doubles from 3 -> 6
- Idle connection timeout ~doubles from 32->60
- Maximum packet timeout is cut in a quarter from 60->15, for more
  retries before triggering an idle timeout. (This will affect
  mid-stream timeouts, but won't affect the connection timeouts we're
  seeing)