chore: pin reth dependencies to a release
Building against a moving target is hard. It's just good practice to
have everyone working against a consistent version of the dependency
anyway. The "version" attribute was basically ignored by cargo.

So those are the theoretical reasons. Here is a practical one:
Basic things are changing about the RPC structs, like the type of the
transaction hashes, even in just the last few weeks.