test: find content should credit the earliest peer
When multiple peers return the content, the result should give
credit to the first peer that responds, not the last.

Also, this bug was not found locally because 10 iterations wasn't enough
to expose the issue the first time. Increasing to 100, which still takes
only 200ms, and was reliably reproducing this issue.