feat: shorter utp timeout, for more reliable query
Shrink the idle timeout from the default of 60 seconds to 5. This comes
up most often in failed connections, which are perpetually idle, and
keep trying until the timeout expires.

During a find-content call, if a uTP transfer is initiated and the
remote peer fails to connect, the default timeout waits for 60 seconds.
The problem is that 60s is also our entire query timeout. So if the utp
connection fails by timeout, then there is no time to attempt downloads
from alternate peers.

By shrinking the timeout to 5 seconds, we leave plenty of room for
contacting other peers for the same content. It still allows us to make
a few attempts at connection (after 1s, then another 1.5s, then another
2.25s).