feat: limit radius to at most 1/8 of MAX
Rationale:
Nodes will not actually store all data. Even if they did, it would be
inefficient to find the data. So we might as well skip the step of
storing and then deleting a bunch of data outside the radius.

Also, state network is using a lot higher CPU than expected. Maybe
tamping the starting radius down a bit at the beginning will shrink CPU
usage during that load-data-in phase.

Also, 10% is the cutoff on glados for considering when a node is
actually storing all the data it offers to store in the radius. So
starting at 12.5% means it is correctly not counted in that set of
filled-up nodes, until it actually shrinks below due to data limits.