docs: major rework of trin book
Introduction is more focused on first-time users. It pushes back some of
the conceptual stuff, tries to explain several things more clearly, and
updates some stale info.

Also, it adds an overall Portal benefits page. It's just a rough first
stab, to have something in place to update as we move forward.