feat: update state:history storage to 1:1
This approach to balancing storage will not last long. We will implement
a more sophisticated approach in the near future.

Rather than trying to guess what the state:history ratio will be, we can
just look at it short term.

If we add in the data from block 21M + 6 months, then the ratio gets to
about 1:1. So technically we are still over-allocating for state right
this moment, but it gives us plenty of runway, and is close enough that
it shouldn't cause major problems.